Improve interface representations of relevant types
