+++ src/Provider/Provider.ts
@@ -1,5 +1,5 @@
 import { TransactionInput, TransactionOutput } from '../Transaction'
-import { Utxo } from '../Wallet/lib'
+import { Utxo } from '../Wallet'
 
 export type SubmitTransaction = (signedTransaction: string) => Promise<boolean>
 export type QueryUtxosByAddress = (addresses: string[]) => Promise<Utxo[]>
