+++ src/Wallet/index.ts
@@ -1,4 +1,6 @@
 import { Wallet } from './Wallet'
 
-export { AddressType, addressDiscoveryWithinBounds } from './lib/address_discovery'
+export { Address, AddressType } from './Address'
+export { Utxo, UtxoWithAddressing } from './Utxo'
+export { addressDiscoveryWithinBounds } from './lib/address_discovery'
 export default Wallet
