+++ src/Wallet/lib/select_inputs.ts
@@ -3,14 +3,7 @@ import { InsufficientValueInUtxosForSelection, MaximumTransactionInputsExceeded 
 import { getRandomBytesForEnvironmentAsHex } from '../../lib/bindings'
 import { MAX_TRANSACTION_INPUTS } from '../config'
 import { InputSelectionAlgorithm } from '../Wallet'
-
-export interface Utxo {
-  address: string
-  value: string
-  hash: string
-}
-
-export type UtxoWithAddressing = Utxo & TransactionInput['addressing']
+import { UtxoWithAddressing } from '..'
 
 export interface TransactionSelection {
   inputs: TransactionInput[]
